refactor(examples): Make "Textures" example upper-case
